Changed how a couple parameters work in main.
The parameter aggDistance now takes a number between 0 and 1.  Later in main, this number is multiplied by gridSize to obtain the actual distance between molecules needed for an aggregate to form.

the two parameters that affect the plot have been replaced with one variable.  if it's true, main returns a plot.